pyproject.toml
[project]
name = "bookmanagment"
version = "0.1.0"
description = ""
authors = [
    {name = "shkrobik2017",email = "shkrobik2017@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "tortoise-orm[asyncpg] (>=0.24.2,<0.25.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "aerich (>=0.8.2,<0.9.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
